Revert "DB 업데이트"
This reverts commit c96dfefa33486a0764938f22881904efd106a688.

+++ server/Client/Win32Project1/Server.cpp
@@ -9,7 +9,7 @@ using namespace std::chrono;
 
 WSADATA Server::wsa;
 SOCKET  Server::sock;
-int	Server::clientID=-1;
+WORD	Server::clientID=-1;
 thread*  Server::receiveThread;
 thread*  Server::updateThread;
 FuncProcess Server::Dispatcher[PACKET_TYPE];
